# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Shopecart E-commerce API Documentation",
        "description": "Documentation des points d'accès (endpoints) de l'API Shopecart.",
        "contact": {
            "email": "contact@shopecart.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Serveur de l'API locale"
        }
    ],
    "paths": {
...
        "/api/orders": {
            "get": {
                "tags": [
                    "Orders"
                ],
                "summary": "Get user's orders",
                "operationId": "2cd7e03021e72d9d799c12dfe179c865",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Order"
                                            }
                                        },
                                        "links": {
                                            "type": "object"
                                        },
                                        "meta": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Orders"
                ],
                "summary": "Create a new order",
                "operationId": "d690d6834318e9b064df441e854de8b9",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "customer_name",
                                    "customer_email",
                                    "customer_phone",
                                    "shipping_address",
                                    "shipping_city",
                                    "shipping_zipcode",
                                    "shipping_country",
                                    "payment_method"
                                ],
                                "properties": {
                                    "customer_name": {
                                        "type": "string",
                                        "example": "John"
                                    },
                                    "customer_email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "john@example.com"
                                    },
                                    "customer_phone": {
                                        "type": "string",
                                        "example": "+1234567890"
                                    },
                                    "shipping_address": {
                                        "type": "string",
                                        "example": "123 Main St"
                                    },
                                    "shipping_city": {
                                        "type": "string",
                                        "example": "New York"
                                    },
                                    "shipping_zipcode": {
                                        "type": "string",
                                        "example": "10001"
                                    },
                                    "shipping_country": {
                                        "type": "string",
                                        "example": "USA"
                                    },
                                    "billing_address": {
                                        "type": "string",
                                        "example": "123 Main St"
                                    },
                                    "billing_city": {
                                        "type": "string",
                                        "example": "New York"
                                    },
                                    "billing_zipcode": {
                                        "type": "string",
                                        "example": "10001"
                                    },
                                    "billing_country": {
                                        "type": "string",
                                        "example": "USA"
                                    },
                                    "payment_method": {
                                        "type": "string",
                                        "example": "credit_card"
                                    },
                                    "notes": {
                                        "type": "string",
                                        "example": "Please deliver in the morning"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Order created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        },
                                        "order": {
                                            "$ref": "#/components/schemas/Order"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/orders/{order}/status": {
            "put": {
                "tags": [
                    "Orders"
                ],
                "summary": "Update the status of a specific order (Management Only)",
                "operationId": "ca851ecbdeb079209abac2e00c4a6359",
                "parameters": [
                    {
                        "name": "order",
                        "in": "path",
                        "description": "ID of the order to update",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "status": {
                                        "description": "New status for the order",
                                        "type": "string",
                                        "enum": [
                                            "PENDING",
                                            "PROCESSING",
                                            "PAID",
                                            "SHIPPED",
                                            "DELIVERED",
                                            "CANCELED",
                                            "FAILED",
                                            "IN_DELIVERY"
                                        ]
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Order status updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Order status updated to SHIPPED"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/Order"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Forbidden - Only management roles can update status"
                    },
                    "404": {
                        "description": "Order not found"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/orders/{id}": {
            "get": {
                "tags": [
                    "Orders"
                ],
                "summary": "Get order details",
                "operationId": "c29621029e70f65247909d3e09031118",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Order ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Order"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Unauthorized"
                    },
                    "404": {
                        "description": "Order not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/orders/my": {
            "get": {
                "tags": [
                    "Orders"
                ],
                "summary": "Get list of orders for the authenticated user (Client Only)",
                "operationId": "bbe2ec3088f531fcb7cb01b1b2c8fa89",
                "parameters": [
                    {
                        "name": "status",
                        "in": "query",
                        "description": "Filter by order status",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "User's orders retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/OrderResource"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "404": {
                        "description": "No orders found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Order": {
                "title": "Order Model",
                "description": "Modèle de données complet pour une commande client.",
                "properties": {
                    "id": {
                        "description": "ID unique de la commande.",
                        "type": "integer",
                        "format": "int64"
                    },
                    "user_id": {
                        "description": "ID de l'utilisateur qui a passé la commande.",
                        "type": "integer",
                        "format": "int64"
                    },
                    "order_number": {
                        "description": "Numéro unique de la commande (ex: ORD-20251114-abcxyz).",
                        "type": "string"
                    },
                    "status": {
                        "description": "Statut de la commande.",
                        "type": "string",
                        "enum": [
                            "PENDING",
                            "PROCESSING",
                            "PAID",
                            "SHIPPED",
                            "DELIVERED",
                            "CANCELED",
                            "FAILED",
                            "IN_DELIVERY"
                        ]
                    },
                    "subtotal": {
                        "description": "Sous-total des articles.",
                        "type": "number",
                        "format": "float"
                    },
                    "shipping": {
                        "description": "Frais de livraison.",
                        "type": "number",
                        "format": "float"
                    },
                    "tax": {
                        "description": "Taxes appliquées.",
                        "type": "number",
                        "format": "float"
                    },
                    "total": {
                        "description": "Prix total final de la commande.",
                        "type": "number",
                        "format": "float"
                    },
                    "customer_first_name": {
                        "description": "Prénom du client.",
                        "type": "string"
                    },
                    "customer_last_name": {
                        "description": "Nom de famille du client.",
                        "type": "string"
                    },
                    "customer_email": {
                        "description": "Email du client.",
                        "type": "string",
                        "format": "email"
                    },
                    "customer_phone": {
                        "description": "Téléphone du client.",
                        "type": "string"
                    },
                    "shipping_address": {
                        "description": "Adresse de livraison.",
                        "type": "string"
                    },
                    "shipping_city": {
                        "description": "Ville de livraison.",
                        "type": "string"
                    },
                    "shipping_zipcode": {
                        "description": "Code postal de livraison.",
                        "type": "string"
                    },
                    "shipping_country": {
                        "description": "Pays de livraison.",
                        "type": "string"
                    },
                    "payment_method": {
                        "description": "Méthode de paiement utilisée.",
                        "type": "string"
                    },
                    "items": {
                        "description": "Articles inclus dans la commande (relation OrderItem).",
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/OrderItem"
                        }
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
...
        },
...
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Entrez le jeton Bearer obtenu après la connexion.",
                "bearerFormat": "Sanctum",
                "scheme": "bearer"
            },
            "sanctum": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
...
}